[push,#85][s]: set file length for rawstore upload from resource object rather than rawstore creds.
This is a bugfix as without it was getting:

Error uploading to rawstore for datapackage.json with code 400 reason
<Error><Code>MalformedPOSTRequest</Code><Message>The body of your POST request is not well-formed multipart/form-data.</Message> ...

We could no longer use rawstore creds because of this upstream change in bitstore: 03415ed7c51bc2cb8a978e9f7f6b944f3bc54155 that removed returning file length as part of the creds. (With that change in turn due to our changes here in f35273c67ebe41fa39baf7ebd457c5f3a72f8768 where we switched to "virtual" datapackage.json resource - that meant the resource no longer had a name property).