[sentry][m]: promisified sending reports to Sentry - refs #329
This is necessary as we want to process.exit once report is sent but we also don't want rest of the code to run if an error occured. To make it work, we've decided to promisify the report sending part of the 'handleError' function and await until it finishes from the caller.
This commit also includes changes in `/bin/` modules to work with refactored `handleError` function.